app/FakeNewsDetectorAPI/requirements.txt
django==4.2.3
djangorestframework==3.14.0
django-filter==24.2
django-cors-headers==4.4.0
psycopg2-binary
scikit-learn==1.5.0
requests==2.32.3
coreapi
